new components
new components for each part of the table
