--- src/components/FlavorItem.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const FlavorItem = ({ flavor }) => {
-    return (
-        <div className="bg-gray-300 h-12 flex items-center justify-center">
-            {flavor}
-        </div>
-    )
-}
