--- src/components/ModalPerson.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const ModalPerson = () => {
-    return (
-        <>
-            
-        </>
-    )
-}
