--- src/components/PersonItem.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const PersonItem = () => {
-    return (
-        <>
-
-        </>
-    )
-}
