--- src/components/TotalFlavorsItem.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const TotalFlavorsItem = ({amount}) => {
-    return (
-        <div className="bg-gray-300 h-12 w-10 flex items-center justify-center">
-            {amount}
-        </div>
-    )
-}
