+++ src/utils/constants.js
@@ -1,0 +1,27 @@
+export const DATA = {
+    names: {
+        jorge: {
+            name: "Jorge",
+            flavors: {
+                jamonyqueso: 1,
+                carne: 2,
+            },
+        },
+        ariel: {
+            name: "Ariel",
+            flavors: {
+                jamonyqueso: 2,
+                carne: 2,
+            },
+        },
+        lujan: {
+            name: "Lujan",
+            flavors: {},
+        },
+    },
+    flavors: {
+        jamonyqueso: "Jamon y queso",
+        carne: "Carne",
+        pollo: "Pollo",
+    },
+};
